requirements.txt
# numpy
cmake
easydict~=1.9
opencv-python~=4.5.1.48
matplotlib~=3.3.3
scipy~=1.5.4
scikit-image
tqdm
PyYAML
open3d
cython
cachetools
tensorboardX
shapely==2.0.0
einops
timm
rospy -i https://rospypi.github.io/simple/
pyzmq
grpcio
